requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6; python_version < "3.13"
openai==0.27.0
channels==4.0.0
asgiref==3.8.1
twilio==8.13.0
gunicorn
whitenoise
Pillow==10.4.0
